Moved different parsing function to own file to make sections of the parser easier to read
