+++ src/parser/CMakeLists.txt
@@ -1,4 +1,11 @@
-loxmocha_add_library(parser parser.cpp)
+loxmocha_add_library(parser
+    parser.cpp
+    parser-decl.cpp
+    parser-expr.cpp
+    parser-pattern.cpp
+    parser-stmt.cpp
+    parser-type.cpp
+)
 
 target_link_libraries(loxmocha_parser
     PUBLIC
